requirements.txt
ansible-base==2.10.17